requirements.txt
torch==1.10.1
Pillow==8.4.0
matplotlib==3.5.1
torchvision==0.11.2
tqdm==4.63.0